Umiddelbart fungerer det nu at rette informationerne på en kunde så længe man først har valgt kunden og trykket vis.

+++ FoxtrotProject/Model/Customer.cs
@@ -17,7 +17,7 @@ namespace FoxtrotProject.Model
 
         public string ContactPerson { get; set; }
 
-        public double GrossIncome { get; set; }
+        public int GrossIncome { get; set; }
 
         public int CVR { get; set; }
